update tests to support doctrine cache 2.0

+++ tests/TestAnnotationReader.php
@@ -12,8 +12,10 @@ namespace Ynlo\GraphQLBundle\Tests;
 
 use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Annotations\CachedReader;
+use Doctrine\Common\Annotations\PsrCachedReader;
 use Doctrine\Common\Annotations\Reader;
 use Doctrine\Common\Cache\PhpFileCache;
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
 
 /**
  * Create Annotation reader with cache to use in tests
@@ -27,12 +29,18 @@ class TestAnnotationReader
      */
     public static function create(): Reader
     {
-        $cache = new PhpFileCache(sys_get_temp_dir());
-        if (!self::$initialized) {
-            $cache->deleteAll();
-            self::$initialized = true;
+        // doctrine/cache ~1.0
+        if (class_exists('Doctrine\Common\Cache\PhpFileCache')) {
+            $cache = new PhpFileCache(sys_get_temp_dir());
+            if (!self::$initialized) {
+                $cache->deleteAll();
+                self::$initialized = true;
+            }
+
+            return new CachedReader(new AnnotationReader(), $cache);
         }
 
-        return new CachedReader(new AnnotationReader(), $cache);
+        // doctrine/cache ~2.0
+        return new PsrCachedReader(new AnnotationReader(), new ArrayAdapter());
     }
 }
